refactor(bubbler-common): update Bubble display properties to use float types for better precision

+++ bubbler-common-api/src/main/java/me/whereareiam/socialismus/module/bubbler/api/model/bubble/Bubble.java
@@ -35,8 +35,8 @@ public class Bubble {
 	@SuperBuilder(toBuilder = true)
 	public static class Display {
 		private int radius;
-		private int headLineGap;
-		private int lineSpacing;
+		private float headLineGap;
+		private float lineSpacing;
 
 		private int maxLinesCount;
 		private int maxLineWidth;
